+++ h3ron/CHANGES.md
@@ -14,6 +14,7 @@ python extension.
 ### Added
 
 - Add `H3Edge::from_cells` as an alternative way to build edges to `H3Cell::unidirectional_edge_to`.
+- Add `h3ron::iter::continuous_cells_to_edges`.
 
 ### Changed